+++ resip/stack/Uri.cxx
@@ -1377,7 +1377,7 @@ Uri::parseEmbeddedHeaders(ParseBuffer& pb)
       else
       {
          DebugLog(<< "Uri::parseEmbeddedHeaders(" << headerName << ", " << Data(decodedContents, len) << ")");
-         mEmbeddedHeaders->addHeader(Headers::getType(headerName.data(), (int)headerName.size()),
+         mEmbeddedHeaders->addHeader(Headers::getType(headerName.data(), headerName.size()),
                                      headerName.data(), (int)headerName.size(),
                                      decodedContents, len);
       }
